requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
enum34==1.1.6
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysql-connector-python-rf==2.2.2
MySQL-python==1.2.5
pylint==1.7.2
PyMySQL==0.7.11
pytz==2017.2
rapidpro-python==2.1.8
requests==2.18.4
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.22
wrapt==1.10.11
